희라 11/13 주말 작업 건

+++ s202350103/src/main/java/com/choongang/s202350103/model/OrderGift.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class OrderGift {		  // 주문_선물
-	private int    o_order_num;   // 주문번호
+	private long   o_order_num;   // 주문번호
 	private String o_gift_num;    // 선물번호
 	private String o_gift_name;   // 받는사람_이름
 	private String o_gift_ph;     // 받는사람_전화번호
